+++ packages/dust_dart/lib/fp.dart
@@ -20,5 +20,8 @@
 library;
 
 export 'src/fp/option.dart';
+export 'src/fp/option_combine.dart';
+export 'src/fp/option_query.dart';
+export 'src/fp/option_transform.dart';
 export 'src/fp/result.dart';
 export 'src/fp/unit.dart';